+++ src/config/index.ts
@@ -1,3 +1,5 @@
 
-export * from './spawn'
+export * from './args'
 export * from './gen'
+export * from './spawn'
+export * from './common'