+++ src/index.ts
@@ -1,6 +1,9 @@
 
-import { dna, bridge, dpki, genConfig } from './config'
+import ConfigBuilder from './config/builder'
+export const Config = ConfigBuilder
 
-export const Config = { dna, bridge, dpki, genConfig }
 export { Orchestrator } from './orchestrator'
+export { Player } from './player'
+export { Instance } from './instance'
 export * from './middleware'
+export * from './types'