+++ src/util.ts
@@ -5,7 +5,18 @@ import logger from './logger'
 import { ObjectS } from './types';
 
 export const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms))
-export const trace = (x, msg = '{T}') => (console.log(msg, typeof x, x), x)
+export const trace = (x, msg = '{T}') => (console.log(msg, `<${typeof x}>`, x), x)
+export const stringify = x => JSON.stringify(x, null, 2)
+
+export const stripPortFromUrl = url => {
+  const i = url.lastIndexOf(':')
+  const maybePort = url.substring(i + 1)
+  if (maybePort.match(/^\d{1,5}$/)) {
+    return url.substring(0, i)
+  } else {
+    throw new Error(`No port found in string "${url}"`)
+  }
+}
 
 // from https://hackernoon.com/functional-javascript-resolving-promises-sequentially-7aac18c4431e
 export function promiseSerialArray<T>(promises: Array<Promise<T>>): Promise<Array<T>> {
@@ -22,6 +33,7 @@ export function promiseSerialObject<T>(promises: ObjectS<Promise<T>>): Promise<O
     Promise.resolve({}))
 }
 
+/** @deprecated */
 export const downloadFile = async ({ url, path, overwrite }: { url: string, path: string, overwrite?: boolean }): Promise<void> => {
   if (overwrite) {
     await _downloadFile({ url, path })
@@ -31,6 +43,7 @@ export const downloadFile = async ({ url, path, overwrite }: { url: string, path
   }
 }
 
+/** @deprecated */
 const _downloadFile = async ({ url, path }: { url: string, path: string }): Promise<void> => {
   const response = await axios.request({
     url: url,