+++ test/unit/test-api.ts
@@ -3,25 +3,28 @@ import * as tape from 'tape'
 import tapeP from 'tape-promise'
 const test = tapeP(tape)
 
-import { Orchestrator } from '../../src';
 import { ScenarioApi } from '../../src/api'
-import { configSugared } from './test-config';
-import * as C from '../../src/config';
+import { instancesDry } from './test-gen';
+import Builder from '../../src/config/builder';
 import * as Gen from '../../src/config/gen';
-import { GenConfigArgs } from '../../src/types';
+import { ConfigSeedArgs } from '../../src/types';
+import { trace } from '../../src/util';
+import { testOrchestrator } from '../common';
 
 test('API detects duplicate agent IDs', async t => {
   const stubGetDnaHash = sinon.stub(Gen, 'getDnaHash').resolves('fakehash')
-  const orchestrator = new Orchestrator()
+  const orchestrator = testOrchestrator()
   const api = new ScenarioApi("description", orchestrator, "uuid")
   const args = {
-    conductorName: 'same',
+    playerName: 'same',
     uuid: 'also-same',
-  } as GenConfigArgs
+  } as ConfigSeedArgs
   await t.rejects(
     api.players({
-      alice: C.desugarConfig(args, configSugared),
-      bob: C.desugarConfig(args, configSugared)
+      local: {
+        alice: Builder.gen(instancesDry),
+        bob: Builder.gen(instancesDry)
+      }
     }),
     /There are 2 non-unique test agent names specified/
   )